pyproject.toml
[project]
name = "rmidi"
version = "0.1.0"
description = "MIDI 2.0 Patch Selection Application"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.23.2",
    "pydantic>=2.4.2",
    "python-rtmidi>=1.5.5",
    "mido>=1.3.0",
    "pyqt6>=6.5.2",
    "httpx>=0.25.0",
]
